refactor: Deleted viewModel in MemoListItem
* Moved styling logics to UIStyle from MemoViewModel
